Verified that the Driverlinx, U500, PMC MCAPI, and Ortec UMCBI drivers compile, link, and run on Win32.  Not tested with the actual hardware though, since I do not have it here.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@169 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/i_umcbi.c
@@ -378,7 +378,8 @@ mxi_umcbi_write( MX_GENERIC *generic, void *buffer, size_t count )
 }
 
 MX_EXPORT mx_status_type
-mxi_umcbi_num_input_bytes_available( MX_GENERIC *generic, int *input_available )
+mxi_umcbi_num_input_bytes_available( MX_GENERIC *generic,
+				unsigned long *input_available )
 {
 	static const char fname[] = "mxi_umcbi_num_input_bytes_available()";
